+++ eyes.sdk.core/src/test/java/com/applitools/connectivity/MockServerConnector.java
@@ -108,8 +108,7 @@ public class MockServerConnector extends ServerConnector {
     }
 
     @Override
-    public Map<String, DeviceSize> getDevicesSizes(String path)
-    {
+    public Map<String, DeviceSize> getDevicesSizes(String path) {
         return new HashMap<>();
     }