+++ eyes.selenium.java/src/test/java/com/applitools/eyes/selenium/rendering/TestVisualGridEyes.java
@@ -6,6 +6,9 @@ import com.applitools.eyes.RectangleSize;
 import com.applitools.eyes.config.Configuration;
 import com.applitools.eyes.config.ConfigurationProvider;
 import com.applitools.eyes.selenium.wrappers.EyesSeleniumDriver;
+import com.applitools.eyes.visualgrid.model.DeviceSize;
+import com.applitools.eyes.visualgrid.model.IosDeviceInfo;
+import com.applitools.eyes.visualgrid.model.IosDeviceName;
 import com.applitools.eyes.visualgrid.services.RunnerOptions;
 import com.applitools.eyes.visualgrid.services.VisualGridRunner;
 import com.applitools.utils.GeneralUtils;
@@ -14,6 +17,9 @@ import org.openqa.selenium.remote.RemoteWebDriver;
 import org.openqa.selenium.remote.SessionId;
 import org.testng.Assert;
 import org.testng.annotations.Test;
+
+import java.util.Map;
+import java.util.concurrent.atomic.AtomicReference;
 
 import static org.mockito.Mockito.*;
 
@@ -88,10 +94,19 @@ public class TestVisualGridEyes {
         configuration2.setApiKey(expectedApiKey);
         configuration2.setServerUrl(expectedServerUrl);
         configuration2.setProxy(expectedProxy);
+        configuration2.addBrowser(new IosDeviceInfo(IosDeviceName.iPad_7));
         runner = new VisualGridRunner();
         eyes = spy(new VisualGridEyes(runner, configurationProvider2));
         doNothing().when(eyes).setViewportSize(ArgumentMatchers.<EyesSeleniumDriver>any());
-        eyes.setServerConnector(new MockServerConnector());
+        final AtomicReference<String> actualApiKey = new AtomicReference<>();
+        eyes.setServerConnector(new MockServerConnector() {
+            @Override
+            public Map<String, DeviceSize> getDevicesSizes(String path) {
+                actualApiKey.set(getApiKey());
+                return super.getDevicesSizes(path);
+            }
+        });
+
         eyes.open(driver, "app", "test", new RectangleSize(800, 800));
         Assert.assertEquals(runner.getApiKey(), expectedApiKey);
         Assert.assertEquals(eyes.getApiKey(), expectedApiKey);
@@ -102,5 +117,6 @@ public class TestVisualGridEyes {
         Assert.assertEquals(runner.getProxy(), expectedProxy);
         Assert.assertEquals(eyes.getProxy(), expectedProxy);
         Assert.assertEquals(eyes.testList.values().iterator().next().getProxy(), expectedProxy);
+        Assert.assertEquals(actualApiKey.get(), expectedApiKey);
     }
 }